debug

+++ daemon/panels.go
@@ -33,29 +33,40 @@ func (d *Daemon) sweepPanels() {
 	batch := &pgx.Batch{}
 
 	for guildId, panelCount := range guilds {
-		tier, err := d.premium.GetTier(guildId)
+		// get tier from patreon
+		tier, err := d.patreon.GetTier(guildId)
 		if err != nil {
 			sentry.Error(err)
 			continue
 		}
 
 		if tier < premium.Premium {
-			query := `
-DELETE FROM
-	panels
-WHERE
-	"message_id" IN (	
-		SELECT
-			"message_id"
-		FROM
-			panels
-		WHERE
-			"guild_id" = $1
-		LIMIT $2
-	)
-;
-`
-			batch.Queue(query, guildId, panelCount - 1)
+			// if not premium on patreon, we should check for a premium key
+			hasKey, err := d.db.PremiumGuilds.IsPremium(guildId)
+			if err != nil {
+				sentry.Error(err)
+				continue
+			}
+
+			if !hasKey {
+				fmt.Printf("guild %d is not patreon anymore! panel count: %d\n", guildId, panelCount)
+				/*query := `
+				DELETE FROM
+					panels
+				WHERE
+					"message_id" IN (
+						SELECT
+							"message_id"
+						FROM
+							panels
+						WHERE
+							"guild_id" = $1
+						LIMIT $2
+					)
+				;
+				`
+							batch.Queue(query, guildId, panelCount - 1)*/
+			}
 		}
 	}
